Split apart massive selection.rs module
