--- emergence_lib/src/player_interaction/selection.rs
@@ -1,688 +1,0 @@
-//! Selecting tiles to be built on, inspected or modified
-
-use crate::{
-    self as emergence_lib,
-    structures::{commands::StructureCommandsExt, ghost::Ghost},
-};
-use bevy::{
-    prelude::*,
-    utils::{HashMap, HashSet},
-};
-use emergence_macros::IterableEnum;
-
-use hexx::{shapes::hexagon, Hex};
-use leafwing_input_manager::{
-    prelude::{ActionState, InputManagerPlugin, InputMap},
-    user_input::{Modifier, UserInput},
-    Actionlike,
-};
-
-use crate::{
-    asset_management::terrain::TerrainHandles,
-    simulation::geometry::{MapGeometry, TilePos},
-    structures::StructureId,
-    terrain::Terrain,
-};
-
-use super::{cursor::CursorPos, InteractionSystem};
-
-/// Actions that can be used to select tiles.
-///
-/// If a tile is not selected, it will be added to the selection.
-/// If it is already selected, it will be removed from the selection.
-#[derive(Actionlike, Clone, Debug)]
-pub(crate) enum SelectionAction {
-    /// Selects a tile or group of tiles.
-    Select,
-    /// Deselects a tile or group of tiles.
-    Deselect,
-    /// Modifies the selection / deselection to be sequential.
-    Multiple,
-    /// Modifies the selection to cover a hexagonal area.
-    Area,
-    /// Modifies the selection to cover a line between the start and end of the selection.
-    Line,
-    /// Selects the structure on the tile under the player's cursor.
-    ///
-    /// If there is no structure there, the player's selection is cleared.
-    Pipette,
-    /// Sets the zoning of all currently selected tiles to the currently selected structure.
-    ///
-    /// If no structure is selected, any zoning will be removed.
-    Zone,
-    /// Sets the zoning of all currently selected tiles to [`Zoning::None`].
-    ///
-    /// If no structure is selected, any zoning will be removed.
-    ClearZoning,
-    /// Removes all structures from the clipboard.
-    ClearClipboard,
-    /// Rotates the contents of the clipboard clockwise.
-    RotateClipboardRight,
-    /// Rotates the conents of the clipboard counterclockwise.
-    RotateClipboardLeft,
-}
-
-impl SelectionAction {
-    /// The default keybindings for mouse and keyboard.
-    fn kbm_binding(&self) -> UserInput {
-        match self {
-            SelectionAction::Select => MouseButton::Left.into(),
-            SelectionAction::Deselect => MouseButton::Right.into(),
-            SelectionAction::Multiple => Modifier::Shift.into(),
-            SelectionAction::Area => Modifier::Control.into(),
-            SelectionAction::Line => Modifier::Alt.into(),
-            SelectionAction::Pipette => KeyCode::Q.into(),
-            SelectionAction::Zone => KeyCode::Space.into(),
-            SelectionAction::ClearZoning => KeyCode::Back.into(),
-            SelectionAction::ClearClipboard => KeyCode::Escape.into(),
-            SelectionAction::RotateClipboardRight => KeyCode::R.into(),
-            SelectionAction::RotateClipboardLeft => {
-                UserInput::modified(Modifier::Shift, KeyCode::R)
-            }
-        }
-    }
-
-    /// The default key bindings
-    fn default_input_map() -> InputMap<SelectionAction> {
-        let mut input_map = InputMap::default();
-
-        for variant in SelectionAction::variants() {
-            input_map.insert(variant.kbm_binding(), variant);
-        }
-        input_map
-    }
-}
-
-/// The zoning of a given tile, which specifies which structure *should* be built there.
-#[derive(Component, PartialEq, Eq, Clone, Debug)]
-pub(crate) enum Zoning {
-    /// The provided structure should be built on this tile.
-    Structure(StructureId),
-    /// No zoning is set.
-    None,
-    /// Zoning is set to keep the tile clear.
-    Clear,
-}
-
-/// How a given object is being interacted with by the player.
-#[derive(PartialEq, Eq, Hash, Clone, Debug, IterableEnum)]
-pub(crate) enum ObjectInteraction {
-    /// Currently in the selection.
-    Selected,
-    /// Hovered over with the cursor.
-    Hovered,
-    /// Hovered over and simultaneously selected.
-    ///
-    /// This exists to allow easy visual distinction of this state,
-    /// and should include visual elements of both.
-    ///
-    // TODO: this is silly and probably shouldn't exist, but we're using colors for everything for now so...
-    // Tracked in https://github.com/Leafwing-Studios/Emergence/issues/263
-    HoveredAndSelected,
-}
-
-impl ObjectInteraction {
-    /// The material used by objects that are being interacted with.
-    pub(crate) fn material(&self) -> StandardMaterial {
-        let base_color = match self {
-            ObjectInteraction::Selected => Color::DARK_GREEN,
-            ObjectInteraction::Hovered => Color::YELLOW,
-            ObjectInteraction::HoveredAndSelected => Color::YELLOW_GREEN,
-        };
-
-        StandardMaterial {
-            base_color,
-            ..Default::default()
-        }
-    }
-}
-
-/// The set of tiles that is currently selected
-#[derive(Resource, Debug, Default, Clone)]
-pub struct SelectedTiles {
-    /// Actively selected tiles
-    selected: HashSet<TilePos>,
-    /// Tiles that are hovered over
-    hovered: HashSet<TilePos>,
-}
-
-impl SelectedTiles {
-    /// Selects a single tile
-    fn add_tile(&mut self, tile_pos: TilePos) {
-        self.selected.insert(tile_pos);
-    }
-
-    /// Deselects a single tile
-    fn remove_tile(&mut self, tile_pos: TilePos) {
-        self.selected.remove(&tile_pos);
-    }
-
-    /// Selects a hexagon of tiles.
-    fn select_hexagon(&mut self, center: TilePos, radius: u32, select: bool) {
-        let hex_coord = hexagon(center.hex, radius);
-
-        for hex in hex_coord {
-            let target_pos = TilePos { hex };
-            // Selection may have overflowed map
-            match select {
-                true => self.add_tile(target_pos),
-                false => self.remove_tile(target_pos),
-            }
-        }
-    }
-
-    /// Clears the set of selected tiles.
-    fn clear_selection(&mut self) {
-        self.selected.clear();
-    }
-
-    /// Are any tiles selected?
-    fn is_empty(&self) -> bool {
-        self.selected.is_empty()
-    }
-
-    /// Is the given tile in the selection?
-    #[cfg(test)]
-    fn contains_tile(&self, tile_pos: TilePos) -> bool {
-        self.selected.contains(&tile_pos)
-    }
-}
-
-/// All tile, structure and unit selection logic and graphics
-pub(super) struct SelectionPlugin;
-
-impl Plugin for SelectionPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<SelectedTiles>()
-            .init_resource::<ActionState<SelectionAction>>()
-            .init_resource::<Clipboard>()
-            .init_resource::<AreaSelection>()
-            .init_resource::<LineSelection>()
-            .insert_resource(SelectionAction::default_input_map())
-            .add_plugin(InputManagerPlugin::<SelectionAction>::default())
-            .add_system(
-                select_tiles
-                    .label(InteractionSystem::SelectTiles)
-                    .after(InteractionSystem::ComputeCursorPos),
-            )
-            .add_system(
-                copy_selection
-                    .label(InteractionSystem::SetClipboard)
-                    .after(InteractionSystem::ComputeCursorPos)
-                    .after(InteractionSystem::SelectTiles),
-            )
-            .add_system(
-                set_zoning
-                    .label(InteractionSystem::ApplyZoning)
-                    .after(InteractionSystem::SelectTiles)
-                    .after(InteractionSystem::SetClipboard),
-            )
-            .add_system(
-                act_on_zoning
-                    .label(InteractionSystem::ManageGhosts)
-                    .after(InteractionSystem::ApplyZoning),
-            )
-            .add_system(
-                display_selection
-                    .label(InteractionSystem::ManageGhosts)
-                    .after(InteractionSystem::SetClipboard),
-            )
-            .add_system(
-                display_tile_interactions
-                    .after(InteractionSystem::SelectTiles)
-                    .after(InteractionSystem::ComputeCursorPos),
-            );
-    }
-}
-
-/// The state needed by [`SelectionAction::Area`].
-#[derive(Resource)]
-struct AreaSelection {
-    /// The central tile, where the area selection began.
-    center: Option<TilePos>,
-    /// The radius of the selection.
-    radius: u32,
-    /// The tiles selected at the start of this action.
-    initial_selection: Option<SelectedTiles>,
-}
-
-impl AreaSelection {
-    /// Set things up to start a line selection action.
-    fn begin(&mut self, selected_tiles: &SelectedTiles, cursor_pos: TilePos) {
-        self.center = Some(cursor_pos);
-        self.initial_selection = Some(selected_tiles.clone());
-    }
-
-    /// Clean things up to conclude a line selection action.
-    fn finish(&mut self) {
-        self.center = None;
-        self.initial_selection = None;
-    }
-}
-
-impl Default for AreaSelection {
-    fn default() -> Self {
-        AreaSelection {
-            center: None,
-            radius: 1,
-            initial_selection: None,
-        }
-    }
-}
-
-/// The state needed by [`SelectionAction::Line`].
-#[derive(Resource, Default)]
-struct LineSelection {
-    /// The starting tile, where the line selection began.
-    start: Option<TilePos>,
-    /// The tiles selected at the start of this action.
-    initial_selection: Option<SelectedTiles>,
-}
-
-impl LineSelection {
-    /// Set things up to start a line selection action.
-    fn begin(&mut self, selected_tiles: &SelectedTiles, cursor_pos: TilePos) {
-        self.start = Some(cursor_pos);
-        self.initial_selection = Some(selected_tiles.clone());
-    }
-
-    /// Clean things up to conclude a line selection action.
-    fn finish(&mut self) {
-        self.start = None;
-        self.initial_selection = None;
-    }
-
-    /// Computes the set of hexagons between `self.start` and `end`, with a thickness determnind by `radius`.
-    fn draw_line(&self, end: TilePos, radius: u32) -> HashSet<TilePos> {
-        let start = self.start.unwrap();
-        let line = start.line_to(end.hex);
-        let mut tiles = HashSet::<TilePos>::new();
-
-        for line_hex in line {
-            let hexagon = hexagon(line_hex, radius);
-            for hex in hexagon {
-                tiles.insert(TilePos { hex });
-            }
-        }
-        tiles
-    }
-}
-
-/// Integrates user input into tile selection actions to let other systems handle what happens to a selected tile
-#[allow(clippy::too_many_arguments)]
-fn select_tiles(
-    cursor: Res<CursorPos>,
-    mut selected_tiles: ResMut<SelectedTiles>,
-    actions: Res<ActionState<SelectionAction>>,
-    mut area_selection: ResMut<AreaSelection>,
-    mut line_selection: ResMut<LineSelection>,
-) {
-    if let Some(cursor_pos) = cursor.maybe_tile_pos() {
-        let select = actions.pressed(SelectionAction::Select);
-        let deselect = actions.pressed(SelectionAction::Deselect);
-
-        let multiple = actions.pressed(SelectionAction::Multiple);
-        let area = actions.pressed(SelectionAction::Area);
-        let line = actions.pressed(SelectionAction::Line);
-        let simple_area = area & !multiple & !line;
-        let simple_deselect = deselect & !area & !multiple & !line;
-
-        // Cache the starting state to make selections reversible
-        if simple_area & area_selection.initial_selection.is_none() {
-            area_selection.begin(&selected_tiles, cursor_pos);
-        }
-
-        if line & line_selection.initial_selection.is_none() {
-            line_selection.begin(&selected_tiles, cursor_pos);
-        }
-
-        // Clean up state from area and line selections
-        if !simple_area {
-            area_selection.finish();
-        }
-
-        if !line {
-            line_selection.finish();
-        }
-
-        // Compute the center and radius
-        let (center, radius) = if area {
-            let center = if !simple_area {
-                cursor_pos
-            } else {
-                area_selection.center.unwrap()
-            };
-
-            if simple_area {
-                area_selection.radius = cursor_pos.unsigned_distance_to(center.hex);
-            }
-
-            (center, area_selection.radius)
-        } else {
-            (cursor_pos, 0)
-        };
-
-        // Record which tiles should have the "hovered" effect
-        selected_tiles.hovered.clear();
-        if simple_area {
-            selected_tiles.hovered.insert(center);
-            let ring = center.hex.ring(radius);
-            for hex in ring {
-                selected_tiles.hovered.insert(TilePos { hex });
-            }
-        } else if line {
-            let line_hexes = line_selection.draw_line(cursor_pos, radius);
-            selected_tiles.hovered.extend(line_hexes);
-        } else {
-            selected_tiles.hovered.insert(cursor_pos);
-        }
-
-        // Don't attempt to handle conflicting inputs.
-        if select & deselect {
-            return;
-        }
-
-        // Clear the selection
-        if simple_deselect | (select & !multiple) {
-            selected_tiles.clear_selection()
-        }
-
-        // Actually select tiles
-        if line {
-            if actions.just_released(SelectionAction::Select) {
-                let line_hexes = line_selection.draw_line(cursor_pos, radius);
-                selected_tiles.selected.extend(line_hexes);
-                line_selection.start = Some(cursor_pos);
-            } else if actions.just_released(SelectionAction::Deselect) {
-                let line_hexes = line_selection.draw_line(cursor_pos, radius);
-                for tile_pos in line_hexes {
-                    selected_tiles.selected.remove(&tile_pos);
-                }
-                line_selection.start = Some(cursor_pos);
-            }
-        } else {
-            if select {
-                selected_tiles.select_hexagon(center, radius, true);
-            }
-
-            if deselect {
-                selected_tiles.select_hexagon(center, radius, false);
-            }
-        }
-    }
-}
-
-/// Shows which tiles are being hovered and selected.
-fn display_tile_interactions(
-    selected_tiles: Res<SelectedTiles>,
-    mut terrain_query: Query<(&mut Handle<StandardMaterial>, &Terrain, &TilePos)>,
-    materials: Res<TerrainHandles>,
-) {
-    if selected_tiles.is_changed() {
-        // PERF: We should probably avoid a linear scan over all tiles here
-        for (mut material, terrain, &tile_pos) in terrain_query.iter_mut() {
-            let hovered = selected_tiles.hovered.contains(&tile_pos);
-            let selected = selected_tiles.selected.contains(&tile_pos);
-
-            *material = materials.get_material(terrain, hovered, selected);
-        }
-    }
-}
-
-/// Stores a selection to copy and paste.
-#[derive(Default, Resource, Debug, Deref, DerefMut)]
-struct Clipboard {
-    /// The internal map of structures.
-    contents: HashMap<TilePos, StructureId>,
-}
-
-impl Clipboard {
-    /// Normalizes the positions of the items on the clipboard.
-    ///
-    /// Centers relative to the median selected tile position.
-    /// Each axis is computed independently.
-    fn normalize_positions(&mut self) {
-        if self.is_empty() {
-            return;
-        }
-
-        let mut x_vec = Vec::from_iter(self.keys().map(|tile_pos| tile_pos.x));
-        let mut y_vec = Vec::from_iter(self.keys().map(|tile_pos| tile_pos.y));
-
-        x_vec.sort_unstable();
-        y_vec.sort_unstable();
-
-        let mid = self.len() / 2;
-        let center = TilePos {
-            hex: Hex {
-                x: x_vec[mid],
-                y: y_vec[mid],
-            },
-        };
-
-        let mut new_map = HashMap::with_capacity(self.capacity());
-
-        for (tile_pos, id) in self.iter() {
-            let new_tile_pos = *tile_pos - center;
-            // PERF: eh maybe we can safe a clone by using remove?
-            new_map.insert(new_tile_pos, id.clone());
-        }
-
-        self.contents = new_map;
-    }
-
-    /// Apply a tile-position shift to the items on the clipboard.
-    ///
-    /// Used to place items in the correct location relative to the cursor.
-    fn offset_positions(&self, origin: TilePos) -> Vec<(TilePos, StructureId)> {
-        self.iter()
-            .map(|(k, v)| ((*k + origin), v.clone()))
-            .collect()
-    }
-}
-
-/// Copies the selected structure(s) to the clipboard, to be placed later.
-///
-/// This system also handles the "pipette" functionality.
-fn copy_selection(
-    cursor: Res<CursorPos>,
-    actions: Res<ActionState<SelectionAction>>,
-    mut clipboard: ResMut<Clipboard>,
-    selected_tiles: Res<SelectedTiles>,
-    structure_query: Query<&StructureId>,
-    map_geometry: Res<MapGeometry>,
-) {
-    if actions.pressed(SelectionAction::ClearClipboard) {
-        clipboard.clear();
-        // Don't try to clear and set the clipboard on the same frame.
-        return;
-    }
-
-    if let Some(cursor_tile_pos) = cursor.maybe_tile_pos() {
-        if actions.just_pressed(SelectionAction::Pipette) {
-            // We want to replace our selection, rather than add to it
-            clipboard.clear();
-
-            // If there is no selection, just grab whatever's under the cursor
-            if selected_tiles.is_empty() {
-                if let Some(structure_entity) = map_geometry.structure_index.get(&cursor_tile_pos) {
-                    let structure_id = structure_query.get(*structure_entity).unwrap();
-                    clipboard.insert(TilePos::default(), structure_id.clone());
-                }
-            } else {
-                for selected_tile_pos in selected_tiles.selected.iter() {
-                    if let Some(structure_entity) =
-                        map_geometry.structure_index.get(selected_tile_pos)
-                    {
-                        let structure_id = structure_query.get(*structure_entity).unwrap();
-                        clipboard.insert(*selected_tile_pos, structure_id.clone());
-                    }
-                }
-                clipboard.normalize_positions();
-            }
-        }
-    }
-}
-
-/// Applies zoning to an area, causing structures to be created (or removed) there.
-///
-/// This system also handles the "paste" functionality.
-fn set_zoning(
-    cursor: Res<CursorPos>,
-    actions: Res<ActionState<SelectionAction>>,
-    clipboard: Res<Clipboard>,
-    mut terrain_query: Query<&mut Zoning, With<Terrain>>,
-    selected_tiles: Res<SelectedTiles>,
-    map_geometry: Res<MapGeometry>,
-) {
-    if let Some(cursor_tile_pos) = cursor.maybe_tile_pos() {
-        let relevant_terrain_entities: Vec<Entity> = if selected_tiles.is_empty() {
-            vec![*map_geometry.terrain_index.get(&cursor_tile_pos).unwrap()]
-        } else {
-            selected_tiles
-                .selected
-                .iter()
-                .map(|tile_pos| *map_geometry.terrain_index.get(tile_pos).unwrap())
-                .collect()
-        };
-
-        // Explicitly clear the selection
-        if actions.pressed(SelectionAction::ClearZoning) {
-            for terrain_entity in relevant_terrain_entities {
-                let mut zoning = terrain_query.get_mut(terrain_entity).unwrap();
-                *zoning = Zoning::Clear;
-            }
-
-            // Don't try to clear and zone in the same frame
-            return;
-        }
-
-        // Apply zoning
-        if actions.pressed(SelectionAction::Zone) {
-            if clipboard.is_empty() {
-                // Clear zoning
-                for terrain_entity in relevant_terrain_entities {
-                    let mut zoning = terrain_query.get_mut(terrain_entity).unwrap();
-                    *zoning = Zoning::Clear;
-                }
-            // Zone using the single selected structure
-            } else if clipboard.len() == 1 {
-                let structure_id = clipboard.values().next().unwrap();
-                for terrain_entity in relevant_terrain_entities {
-                    let mut zoning = terrain_query.get_mut(terrain_entity).unwrap();
-                    *zoning = Zoning::Structure(structure_id.clone());
-                }
-            // Paste the selection
-            } else {
-                for (tile_pos, structure_id) in clipboard.offset_positions(cursor_tile_pos) {
-                    // Avoid trying to operate on terrain that doesn't exist
-                    if let Some(&terrain_entity) = map_geometry.terrain_index.get(&tile_pos) {
-                        let mut zoning = terrain_query.get_mut(terrain_entity).unwrap();
-                        *zoning = Zoning::Structure(structure_id.clone());
-                    }
-                }
-            }
-        }
-    }
-}
-
-/// Spawn and despawn structures based on their zoning.
-fn act_on_zoning(
-    terrain_query: Query<(&Zoning, &TilePos), (With<Terrain>, Changed<Zoning>)>,
-    mut commands: Commands,
-) {
-    for (zoning, &tile_pos) in terrain_query.iter() {
-        match zoning {
-            Zoning::Structure(id) => commands.spawn_structure(tile_pos, id.clone()),
-            Zoning::None => (), // Do nothing
-            Zoning::Clear => commands.despawn_structure(tile_pos),
-        };
-    }
-}
-
-/// Show the current selection under the cursor
-fn display_selection(
-    clipboard: Res<Clipboard>,
-    cursor_pos: Res<CursorPos>,
-    mut commands: Commands,
-    mut ghost_query: Query<(&TilePos, &mut StructureId), With<Ghost>>,
-) {
-    if let Some(cursor_pos) = cursor_pos.maybe_tile_pos() {
-        let mut desired_ghosts: HashMap<TilePos, StructureId> =
-            HashMap::with_capacity(clipboard.capacity());
-        for (&clipboard_pos, structure_id) in clipboard.iter() {
-            let tile_pos = cursor_pos + clipboard_pos;
-            desired_ghosts.insert(tile_pos, structure_id.clone());
-        }
-
-        // Handle ghosts that already exist
-        for (tile_pos, mut existing_structure_id) in ghost_query.iter_mut() {
-            // Ghost should exist
-            if let Some(desired_structure_id) = desired_ghosts.get(tile_pos) {
-                // Ghost's identity changed
-                if *existing_structure_id != *desired_structure_id {
-                    // TODO: Bevy 0.10, use set_if_neq
-                    *existing_structure_id = desired_structure_id.clone();
-                }
-
-                // This ghost has been handled
-                desired_ghosts.remove(tile_pos);
-            // Ghost should no longer exist
-            } else {
-                commands.despawn_ghost(*tile_pos);
-            }
-        }
-
-        // Handle any remaining new ghosts
-        for (&tile_pos, id) in desired_ghosts.iter() {
-            commands.spawn_ghost(tile_pos, id.clone());
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::SelectedTiles;
-    use crate::simulation::geometry::TilePos;
-
-    #[test]
-    fn simple_selection() {
-        let mut selected_tiles = SelectedTiles::default();
-        let tile_pos = TilePos::default();
-
-        selected_tiles.add_tile(tile_pos);
-        assert!(selected_tiles.contains_tile(tile_pos));
-        assert!(!selected_tiles.is_empty());
-        assert_eq!(selected_tiles.selected.len(), 1);
-
-        selected_tiles.remove_tile(tile_pos);
-        assert!(!selected_tiles.contains_tile(tile_pos));
-        assert!(selected_tiles.is_empty());
-        assert_eq!(selected_tiles.selected.len(), 0);
-    }
-
-    #[test]
-    fn multi_select() {
-        let mut selected_tiles = SelectedTiles::default();
-
-        selected_tiles.add_tile(TilePos::new(1, 1));
-        // Intentionally doubled
-        selected_tiles.add_tile(TilePos::new(1, 1));
-        selected_tiles.add_tile(TilePos::new(2, 2));
-        selected_tiles.add_tile(TilePos::new(3, 3));
-
-        assert_eq!(selected_tiles.selected.len(), 3);
-    }
-
-    #[test]
-    fn clear_selection() {
-        let mut selected_tiles = SelectedTiles::default();
-        selected_tiles.add_tile(TilePos::new(1, 1));
-        selected_tiles.add_tile(TilePos::new(2, 2));
-        selected_tiles.add_tile(TilePos::new(3, 3));
-
-        assert_eq!(selected_tiles.selected.len(), 3);
-        selected_tiles.clear_selection();
-        assert_eq!(selected_tiles.selected.len(), 0);
-    }
-}
